+++ anyway/parsers/rss_sites_parsing_utils.py
@@ -1,10 +1,6 @@
-from anyway.parsers.news_flash_classifiers import classify_ynet
-from ..location_extraction import (
-    manual_filter_location_of_text,
-    geocode_extract,
-    get_db_matching_location,
-    set_accident_resolution,
-)
+from .news_flash_classifiers import classify_ynet
+from .location_extraction import manual_filter_location_of_text, geocode_extract, get_db_matching_location, \
+    set_accident_resolution
 
 
 def extract_geo_features(parsed_item, google_maps_key):
